+++ register-callback/index.js
@@ -1,4 +1,4 @@
-import RegisterCallback from './src/RegisterCallback';
+import RegisterCallback from './src/register-callback';
 
 export {
   RegisterCallback